Переименовывает модули и добавляет модель для открытия комментариев
