+++ js/create-array-miniatures.js
