+++ js/get-big-picture.js
@@ -1,19 +1,12 @@
-import './getMiniatures.js';
+import './get-miniatures.js';
 import { isEscapeKey } from './util.js';
-import { descriptionPhoto } from './createArrayMiniatures.js';
+import { descriptionPhoto } from './create-array-miniatures.js';
+import { openComments } from './open-comments.js';
 
 const bigPictureElement = document.querySelector('.big-picture');
 const bigPictureImg = bigPictureElement.querySelector('.big-picture__img img');
 const likesCount = bigPictureElement.querySelector('.likes-count');
 const caption = bigPictureElement.querySelector('.social__caption');
-
-const socialComments = bigPictureElement.querySelector('.social__comments');
-const socialCommentsTemplate = socialComments.querySelector('.social__comment');
-
-const commentCount = bigPictureElement.querySelector('.social__comment-count');
-const commentShownCount = commentCount.querySelector('.social__comment-shown-count');
-const commentTotalCount = commentCount.querySelector('.social__comment-total-count');
-const commentsLoader = bigPictureElement.querySelector('.comments-loader');
 
 const buttonCancel = document.querySelector('.big-picture__cancel');
 
@@ -43,31 +36,15 @@ const closeBigPicture = () => {
 // Функция для открытия большой картинки
 const openBigPicture = (id) => {
   const currentPhoto = descriptionPhoto.find((photo) => photo.id === Number(id));
-  const socialCommentsFragment = document.createDocumentFragment();
 
   bigPictureImg.src = currentPhoto.url;
   bigPictureImg.alt = currentPhoto.description;
   caption.textContent = currentPhoto.description;
   likesCount.textContent = currentPhoto.likes;
-  // Пока оставила показанные комментраии равные общему количеству комментариев
-  commentShownCount.textContent = currentPhoto.comments.length;
-  commentTotalCount.textContent = currentPhoto.comments.length;
 
-  socialComments.innerHTML = '';
-  currentPhoto.comments.forEach((comment) => {
-    const socialComment = socialCommentsTemplate.cloneNode(true);
-
-    socialComment.querySelector('.social__picture').src = comment.avatar;
-    socialComment.querySelector('.social__picture').alt = comment.name;
-    socialComment.querySelector('.social__text').textContent = comment.message;
-
-    socialCommentsFragment.appendChild(socialComment);
-  });
-  socialComments.appendChild(socialCommentsFragment);
+  openComments(id);
 
   bigPictureElement.classList.remove('hidden');
-  commentCount.classList.add('hidden');
-  commentsLoader.classList.add('hidden');
   document.body.classList.add('modal-open');
 
   buttonCancel.addEventListener('click', onCloseBigPictureClick);
